refactor: data/home 패키지에서 data/list로 패키지 이름 수정
