+++ app/src/main/java/com/depromeet/sloth/data/network/list/LessonInfoResponse.kt
@@ -1,4 +1,4 @@
-package com.depromeet.sloth.data.network.home
+package com.depromeet.sloth.data.network.list
 
 /**
  *   AllLessonResponse
