+++ app/src/main/java/com/depromeet/sloth/data/network/list/LessonRepository.kt
@@ -1,4 +1,4 @@
-package com.depromeet.sloth.data.network.home
+package com.depromeet.sloth.data.network.list
 
 import com.depromeet.sloth.BuildConfig
 import com.depromeet.sloth.data.network.ServiceGenerator
