+++ app/src/main/java/com/depromeet/sloth/ui/list/LessonListAdapter.kt
@@ -16,7 +16,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.depromeet.sloth.R
-import com.depromeet.sloth.data.network.home.LessonInfoResponse
+import com.depromeet.sloth.data.network.list.LessonInfoResponse
 import java.text.DecimalFormat
 import kotlin.math.ceil
