+++ app/src/main/java/com/depromeet/sloth/ui/list/TodayLessonAdapter.kt
@@ -15,7 +15,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.depromeet.sloth.R
-import com.depromeet.sloth.data.network.home.LessonTodayResponse
+import com.depromeet.sloth.data.network.list.LessonTodayResponse
 
 class TodayLessonAdapter(
     private val bodyType: BodyType,
